move  types to Common folder
