+++ client/src/Home/HomePage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Container } from "react-bootstrap";
-import type { Group } from "../App/Types";
+import type { Group } from "../Common/Types";
 import Body from "./Body";
 import Header from "./Header";
